Moving tracking code to mixin for use in Organization and OrganizationOption
