+++ src/sentry/db/models/base.py
@@ -1,10 +1,5 @@
-from copy import copy
-import logging
-
-from bitfield.types import BitHandler
 from django.db import models
 from django.db.models import signals
-from django.db.models.query_utils import DeferredAttribute
 from django.utils import timezone
 
 from .fields.bounded import BoundedBigAutoField
@@ -12,10 +7,6 @@ from .manager import BaseManager
 from .query import update
 
 __all__ = ("BaseModel", "Model", "DefaultFieldsModel", "sane_repr")
-
-UNSAVED = object()
-
-DEFERRED = object()
 
 
 def sane_repr(*attrs):
@@ -42,7 +33,6 @@ class BaseModel(models.Model):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self._update_tracked_data()
 
     def __getstate__(self):
         d = self.__dict__.copy()
@@ -63,57 +53,6 @@ class BaseModel(models.Model):
 
     def __setstate__(self, state):
         self.__dict__.update(state)
-        self._update_tracked_data()
-
-    def __get_field_value(self, field):
-        if isinstance(type(field).__dict__.get(field.attname), DeferredAttribute):
-            return DEFERRED
-        if isinstance(field, models.ForeignKey):
-            return getattr(self, field.column, None)
-        return getattr(self, field.attname, None)
-
-    def _update_tracked_data(self):
-        "Updates a local copy of attributes values"
-        if self.id:
-            data = {}
-            for f in self._meta.fields:
-                # XXX(dcramer): this is how Django determines this (copypasta from Model)
-                if (
-                    isinstance(type(f).__dict__.get(f.attname), DeferredAttribute)
-                    or f.column is None
-                ):
-                    continue
-                try:
-                    v = self.__get_field_value(f)
-                except AttributeError as e:
-                    # this case can come up from pickling
-                    logging.exception(str(e))
-                else:
-                    if isinstance(v, BitHandler):
-                        v = copy(v)
-                    data[f.column] = v
-            self.__data = data
-        else:
-            self.__data = UNSAVED
-
-    def has_changed(self, field_name):
-        "Returns ``True`` if ``field`` has changed since initialization."
-        if self.__data is UNSAVED:
-            return False
-        field = self._meta.get_field(field_name)
-        value = self.__get_field_value(field)
-        if value is DEFERRED:
-            return False
-        return self.__data.get(field_name) != value
-
-    def old_value(self, field_name):
-        "Returns the previous value of ``field``"
-        if self.__data is UNSAVED:
-            return None
-        value = self.__data.get(field_name)
-        if value is DEFERRED:
-            return None
-        return self.__data.get(field_name)
 
 
 class Model(BaseModel):
@@ -142,7 +81,6 @@ def __model_pre_save(instance, **kwargs):
 def __model_post_save(instance, **kwargs):
     if not isinstance(instance, BaseModel):
         return
-    instance._update_tracked_data()
 
 
 def __model_class_prepared(sender, **kwargs):
